# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "H&M Fashion Recommendation API",
    "description": "Personalized fashion recommendations via LightGBM LambdaRank.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/articles/search": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Search articles by name/type",
        "operationId": "search_articles_api_v1_articles_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Name or keyword search",
              "title": "Q"
            },
            "description": "Name or keyword search"
          },
          {
            "name": "product_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleMeta"
                  },
                  "title": "Response Search Articles Api V1 Articles Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{article_id}": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Article metadata by ID",
        "operationId": "get_article_api_v1_articles__article_id__get",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{article_id}/image": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Article image (local if available, placeholder otherwise)",
        "operationId": "get_article_image_api_v1_articles__article_id__image_get",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArticleMeta": {
        "properties": {
          "article_idx": {
            "type": "integer",
            "title": "Article Idx"
          },
          "article_id": {
            "type": "string",
            "title": "Article Id"
          },
          "prod_name": {
            "type": "string",
            "title": "Prod Name"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          },
          "colour": {
            "type": "string",
            "title": "Colour"
          },
          "department": {
            "type": "string",
            "title": "Department"
          }
        },
        "type": "object",
        "required": [
          "article_idx",
          "article_id",
          "prod_name",
          "product_type",
          "colour",
          "department"
        ],
        "title": "ArticleMeta"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}